separate `Orbit.jl` into multiple submodules #124
